add policy for api&web

+++ app/Http/Controllers/Api/TweetController.php
@@ -12,6 +12,8 @@ use Illuminate\Http\Request;
 use App\Models\Tweet;
 // 🔽 追加
 use App\Services\TweetService;
+// 🔽 追加(資料3.27)
+use Illuminate\Support\Facades\Gate;
 
 
 class TweetController extends Controller
@@ -29,6 +31,8 @@ class TweetController extends Controller
      */
     public function index()
     {
+        // 🔽 追加(資料3.27)
+        Gate::authorize('viewAny', Tweet::class);
         // 🔽 編集
         $tweets = $this->tweetService->allTweets();
         return response()->json($tweets);
@@ -40,6 +44,8 @@ class TweetController extends Controller
     // 🔽 編集
     public function store(StoreTweetRequest $request)
     {
+        // 🔽 追加(資料3.27)
+        Gate::authorize('create', Tweet::class);
         // バリデーションは削除
         $tweet = $this->tweetService->createTweet($request);
         return response()->json($tweet, 201);
@@ -50,6 +56,8 @@ class TweetController extends Controller
      */
     public function show(Tweet $tweet)
     {
+        // 🔽 追加(資料3.27)
+        Gate::authorize('view', $tweet);
         return response()->json($tweet);
     }
 
@@ -58,6 +66,8 @@ class TweetController extends Controller
      */
     public function update(UpdateTweetRequest $request, Tweet $tweet)
     {
+        // 🔽 追加(資料3.27)
+        Gate::authorize('update', $tweet);
         // 🔽 編集 + バリデーションは削除
         $updatedTweet = $this->tweetService->updateTweet($request, $tweet);
         return response()->json($updatedTweet);
@@ -68,6 +78,8 @@ class TweetController extends Controller
      */
     public function destroy(Tweet $tweet)
     {
+        // 🔽 追加(資料3.27)
+        Gate::authorize('delete', $tweet);
         // 🔽 編集
         $this->tweetService->deleteTweet($tweet);
         return response()->json(['message' => 'Tweet deleted successfully']);
